Authorization checks for edit/delete posts

+++ controller/feed.js
@@ -129,6 +129,11 @@ exports.updatePost = (req,res,next) => {
                 error.statusCode = 404;
                 throw error;  
             }
+            if(post.creator.toString() !== req.userId){
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
             if(imageUrl !== post.imageUrl){
                clearImage(post.imageUrl); 
             }
@@ -160,7 +165,11 @@ exports.deletePost = (req,res,next) => {
                 error.statusCode = 404;
                 throw error;  
             }
-            //check logged in user
+            if(post.creator.toString() !== req.userId){
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
